+++ value_bloc/test/ListValueBloc_test.dart
@@ -51,12 +51,11 @@ void main() {
           LoadingListValueState(delegate),
           LoadedListValueState(delegate),
           FetchingListValueState(delegate),
-          FetchedListValueState(delegate = delegate.rebuild((b) =>
-              b..pages[FetchScheme(0, 2)] = TestPagesBloc.values.take(2).toBuiltList())),
+          FetchedListValueState(delegate = delegate.rebuild(
+              (b) => b..pages[FetchScheme(0, 2)] = TestPagesBloc.values.take(2).toBuiltList())),
           FetchingListValueState(delegate),
-          FetchedListValueState(delegate = delegate.rebuild((b) => b
-            ..pages[FetchScheme(4, 2)] =
-                TestPagesBloc.values.skip(4).take(2).toBuiltList())),
+          FetchedListValueState(delegate = delegate.rebuild((b) =>
+              b..pages[FetchScheme(4, 2)] = TestPagesBloc.values.skip(4).take(2).toBuiltList())),
           // Refresh -> 6
           FetchingListValueState(delegate.rebuild((b) => b..clearAfterFetch = true)),
           FetchedListValueState(delegate = delegate.rebuild((b) => b
@@ -71,6 +70,47 @@ void main() {
           //   ..pages[FetchScheme(0, 2)] = TestPagesBloc.values.take(2).toBuiltList())),
         ],
       );
+    });
+
+    test("test values without schemes", () {
+      final state = FetchedListValueState(ListValueStateDelegate<int, Object>((b) => b
+        ..clearAfterFetch = false
+        ..pages));
+
+      expect(state.values, BuiltMap<int, int>());
+    });
+
+    test("test values with single scheme", () {
+      final state = FetchedListValueState(ListValueStateDelegate<int, Object>((b) => b
+        ..clearAfterFetch = false
+        ..pages.addAll({
+          FetchScheme(0, 2): BuiltList([0, 1]),
+        })));
+
+      expect(
+          state.values,
+          BuiltMap<int, int>({
+            0: 0,
+            1: 1,
+          }));
+    });
+
+    test("test values with more schemes", () {
+      final state = FetchedListValueState(ListValueStateDelegate<int, Object>((b) => b
+        ..clearAfterFetch = false
+        ..pages.addAll({
+          FetchScheme(0, 2): BuiltList([0, 1]),
+          FetchScheme(2, 2): BuiltList([2, 3]),
+        })));
+
+      expect(
+          state.values,
+          BuiltMap<int, int>({
+            0: 0,
+            1: 1,
+            2: 2,
+            3: 3,
+          }));
     });
   });
 }